+++ localflavor/at/forms.py
@@ -2,8 +2,7 @@
 import re
 
 from django.core.validators import EMPTY_VALUES
-from django.forms import ValidationError
-from django.forms.fields import Field, RegexField, Select
+from django.forms import ValidationError, CharField, RegexField, Select
 from django.utils.translation import gettext_lazy as _
 
 from .at_states import STATE_CHOICES
@@ -33,7 +32,7 @@ class ATStateSelect(Select):
         super().__init__(attrs, choices=STATE_CHOICES)
 
 
-class ATSocialSecurityNumberField(Field):
+class ATSocialSecurityNumberField(CharField):
     """
     Austrian Social Security numbers are composed of a 4 digits and 6 digits field.
 
@@ -53,8 +52,8 @@ class ATSocialSecurityNumberField(Field):
 
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ""
+        if value in self.empty_values:
+            return self.empty_value
         if not re_ssn.search(value):
             raise ValidationError(self.error_messages['invalid'], code='invalid')
         sqnr, date = value.split(" ")