+++ localflavor/ca/forms.py
@@ -2,9 +2,10 @@
 
 import re
 
-from django.core.validators import EMPTY_VALUES
+from django.core.exceptions import ImproperlyConfigured
+
 from django.forms import ValidationError
-from django.forms.fields import CharField, Field, Select
+from django.forms.fields import CharField, Select
 from django.utils.translation import gettext_lazy as _
 from stdnum import luhn
 
@@ -39,13 +40,17 @@ class CAPostalCodeField(CharField):
         return "%s %s" % (m.group(1), m.group(2))
 
 
-class CAProvinceField(Field):
+class CAProvinceField(CharField):
     """
     A form field that validates its input is a Canadian province name or abbreviation.
 
     It normalizes the input to the standard two-leter postal service
     abbreviation for the given province.
     """
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and kwargs["strip"] is False:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
 
     default_error_messages = {
         'invalid': _('Enter a Canadian province or territory.'),
@@ -53,19 +58,14 @@ class CAProvinceField(Field):
 
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
         try:
-            value = value.strip().lower()
-        except AttributeError:
-            pass
-        else:
             # Load data in memory only when it is required, see also #17275
             from .ca_provinces import PROVINCES_NORMALIZED
-            try:
-                return PROVINCES_NORMALIZED[value.strip().lower()]
-            except KeyError:
-                pass
+            return PROVINCES_NORMALIZED[value.lower()]
+        except KeyError:
+            pass
         raise ValidationError(self.error_messages['invalid'], code='invalid')
 
 
@@ -78,7 +78,7 @@ class CAProvinceSelect(Select):
         super().__init__(attrs, choices=PROVINCE_CHOICES)
 
 
-class CASocialInsuranceNumberField(Field):
+class CASocialInsuranceNumberField(CharField):
     """
     A Canadian Social Insurance Number (SIN).
 
@@ -98,8 +98,8 @@ class CASocialInsuranceNumberField(Field):
 
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return self.empty_value
 
         match = re.match(sin_re, value)
         if not match: